+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
 
+- Expose project CA certificate to service secrets: `REDIS_CA_CERT`, `MYSQL_CA_CERT`, etc.
 - Add `KafkaTopic` field `config.local_retention_bytes`, type `integer`: local.retention.bytes value
 - Add `KafkaTopic` field `config.local_retention_ms`, type `integer`: local.retention.ms value
 - Add `KafkaTopic` field `config.remote_storage_enable`, type `boolean`: remote_storage_enable